feat: add slices for evm store
