+++ apps/trading/lib/hooks/use-evm-deposit.ts
@@ -3,16 +3,16 @@ import { useEvmTxStore, type TxDeposit } from './use-evm-tx';
 import { useRef } from 'react';
 
 export const useEvmDeposit = () => {
-  const idRef = useRef<string>(uniqueId());
+  const idRef = useRef<string>();
   const deposit = useEvmTxStore((store) => store.deposit);
-  const transaction = useEvmTxStore((store) => {
-    return store.txs.get(idRef.current);
-  });
+  const txs = useEvmTxStore((store) => store.txs);
+  const data = idRef.current ? txs.get(idRef.current) : undefined;
 
   return {
     write: (config: Parameters<typeof deposit>[1]) => {
+      idRef.current = uniqueId();
       return deposit(idRef.current, config);
     },
-    data: transaction as TxDeposit,
+    data: data as TxDeposit,
   };
 };
